requirements.txt
accelerate
datasets
dill
evaluate
nltk
numpy
pandas
peft
pillow
pyarrow
requests
tqdm
transformers
wandb